# PUBDEV-8971: Speed-up XGBoost scoring on wide datasets