--- a/h2o-extensions/xgboost/src/main/java/hex/tree/xgboost/predict/PredictorFactory.java
+++ b/h2o-extensions/xgboost/src/main/java/hex/tree/xgboost/predict/PredictorFactory.java
@@ -34,7 +34,7 @@ public class PredictorFactory {
     }
   }
 
-  private static boolean unsafeTreesSupported() {
+  static boolean unsafeTreesSupported() {
     // XGBoost Predictor uses LE, we can only use our scoring if the system has the same endianness
     return ByteOrder.nativeOrder().equals(ByteOrder.LITTLE_ENDIAN);
   }