The goal of this PR is to speed-up XGBoost scoring, especially on wide datasets and in cases when only a portion of training columns is used to by the model.

Idea: It is expensive to repeatedly (=for each scoring) convert H2O Frame into float[] rows for Predictor scoring. In this change we only convert the columns that are actually actively used by the model.

Measurements (only measures scoring time during training, time to build trees is ignored)

Case MNIST, 60k rows, 785 cols

    Default settings + score_each_iteration = True

    Old: 15.7s
    New: 9.4s


Syntetic, 300k rows, 800 cols

    Extreme settings - max_depth = 1, ntrees = 10 (this forces a small model - only up to 10 different features can be used by the model)

    Old: 89s
    New: 8.1s (10x+ speed-up)

This is an early preview of the implementation demonstrating the direction I would like to evolve the PR in.