@michalkurka 2023-02-02
@valenad1 @mn-mikke thanks for running the tests and confirming they pass; FIY - this PR is a prerequisite to https://github.com/h2oai/h2o-3/pull/6481