@mn-mikke 2023-02-02
I've re-run the failed tests:

There was one related to external xgboost:
```
versionFromGradle='3.39.0',projectVersion='3.39.0.99999',branch='(HEAD detached at c1b878a)',lastCommitHash='c1b878a95bdaef4814f0ea8bde5c140fbda61719',gitDescribe='jenkins-master-6095-3-gc1b878a',compiledOn='2023-02-02 19:47:57',compiledBy='jenkins'
loading the following test utility modules:  ['/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/tests/_utils.py', '/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/tests/testdir_algos/automl/_automl_utils.py', '/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/tests/testdir_algos/automl/leaderboard/_leaderboard_utils.py']
[2023-02-02 20:14:56] Connect to h2o on IP: 127.0.0.1 PORT: 40006

Not using any auth
Parse progress: |â–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆ| (done) 100%
xgboost Model Build progress: |â–ˆâ–ˆâ–ˆâ–ˆâ–ˆâ–ˆ (failed)
Traceback (most recent call last):
  File "/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/scripts/h2o-py-test-setup.py", line 186, in <module>
    h2o_test_setup(sys.argv)
  File "/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/scripts/h2o-py-test-setup.py", line 181, in h2o_test_setup
    elif _IS_PYUNIT_:    pyunit_utils.pyunit_exec(_TEST_NAME_)
  File "/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/tests/pyunit_utils/utilsPY.py", line 686, in pyunit_exec
    exec(pyunit_c, dict(__name__='__main__', __file__=test_path))  # forcing module name to ensure that the test behaves the same way as when executed using `python my_test.py`
  File "/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/tests/testdir_algos/xgboost/pyunit_xgboost_eval_metric_cv_large.py", line 59, in <module>
    pyunit_utils.standalone_test(test_gam_model_predict)
  File "/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/tests/pyunit_utils/utilsPY.py", line 701, in standalone_test
    test()
  File "/home/jenkins/slave_dir_from_mr-0xc1/workspace/h2o-3-pipeline_PR-6469@2/external-xgboost/h2o-3/h2o-py/tests/testdir_algos/xgboost/pyunit_xgboost_eval_metric_cv_large.py", line 25, in test_gam_model_predict
    xgb_classic.train(y=y, training_frame=loan_df)
  File "/envs/h2o_env_python3.7/lib/python3.7/site-packages/h2o/estimators/estimator_base.py", line 108, in train
    self._train(parms, verbose=verbose)
  File "/envs/h2o_env_python3.7/lib/python3.7/site-packages/h2o/estimators/estimator_base.py", line 200, in _train
    job.poll(poll_updates=self._print_model_scoring_history if verbose else None)
  File "/envs/h2o_env_python3.7/lib/python3.7/site-packages/h2o/job.py", line 91, in poll
    "\n{}".format(self.job_key, self.exception, self.job["stacktrace"]))
OSError: Job with key $03017f000001479cffffffff$_81e08775ec9c46e0ac26bb54b7ff4698 failed with an exception: java.lang.RuntimeException: Error while training XGBoost model
stacktrace: 
java.lang.RuntimeException: Error while training XGBoost model
	at hex.tree.xgboost.XGBoost$XGBoostDriver.buildModel(XGBoost.java:429)
	at hex.tree.xgboost.XGBoost$XGBoostDriver.computeImpl(XGBoost.java:379)
	at hex.ModelBuilder$Driver.compute2(ModelBuilder.java:253)
	at water.H2O$H2OCountedCompleter.compute(H2O.java:1677)
	at jsr166y.CountedCompleter.exec(CountedCompleter.java:468)
	at jsr166y.ForkJoinTask.doExec(ForkJoinTask.java:263)
	at jsr166y.ForkJoinPool$WorkQueue.runTask(ForkJoinPool.java:976)
	at jsr166y.ForkJoinPool.runWorker(ForkJoinPool.java:1479)
	at jsr166y.ForkJoinWorkerThread.run(ForkJoinWorkerThread.java:104)
Caused by: DistributedException from /127.0.0.1:40010: 'Unexpected response (status: HTTP/1.1 500 Server Error).', caused by java.lang.IllegalStateException: Unexpected response (status: HTTP/1.1 500 Server Error).
	at water.MRTask.getResult(MRTask.java:660)
	at water.MRTask.getResult(MRTask.java:670)
	at water.MRTask.doAll(MRTask.java:555)
	at water.MRTask.doAllNodes(MRTask.java:568)
	at hex.tree.xgboost.task.AbstractXGBoostTask.run(AbstractXGBoostTask.java:45)
	at hex.tree.xgboost.exec.RemoteXGBoostExecutor.uploadMatrix(RemoteXGBoostExecutor.java:81)
	at hex.tree.xgboost.exec.RemoteXGBoostExecutor.<init>(RemoteXGBoostExecutor.java:53)
	at hex.tree.xgboost.XGBoost$XGBoostDriver.makeExecutor(XGBoost.java:393)
	at hex.tree.xgboost.XGBoost$XGBoostDriver.buildModel(XGBoost.java:425)
	... 8 more
Caused by: java.lang.IllegalStateException: Unexpected response (status: HTTP/1.1 500 Server Error).
	at hex.tree.xgboost.exec.XGBoostHttpClient.executeRequestAndReturnResponse(XGBoostHttpClient.java:224)
	at hex.tree.xgboost.exec.XGBoostHttpClient.uploadMatrixData(XGBoostHttpClient.java:203)
	at hex.tree.xgboost.task.XGBoostUploadMatrixTask$UploadSparseMatrixFun.map(XGBoostUploadMatrixTask.java:373)
	at water.LocalMR.compute2(LocalMR.java:89)
	... 6 more

Closing connection _sid_a24c at exit
```